refactor(management): finish monitoring decomposition
