+++ cliproxyapi-pro-management/overlay/src/features/monitoring/components/AccountOverviewPanels.tsx
@@ -1,0 +1,493 @@
+import { useState, type ReactNode } from 'react';
+import type { TFunction } from 'i18next';
+import { Card } from '@/components/ui/Card';
+import { IconChevronDown, IconChevronUp, IconRefreshCw } from '@/components/ui/icons';
+import type { QuotaRenderHelpers } from '@/components/quota/QuotaCard';
+import { QuotaProgressBar as AuthFileQuotaProgressBar } from '@/features/authFiles/components/QuotaProgressBar';
+import type { AccountStatusData } from '../accountHealth';
+import {
+  hasUsableQuotaContent,
+  resolveQuotaErrorMessage,
+  type AccountQuotaEntry,
+  type AccountQuotaState,
+} from '../accountQuota';
+import {
+  formatPercent,
+  getAccountHealthTone,
+  type AccountHealthTone,
+} from '../monitoringAnalytics';
+import { MonitoringHealthStatusBar } from './MonitoringHealthStatusBar';
+import {
+  joinUnique,
+  type MonitoringAccountRow,
+} from '../hooks/useMonitoringData';
+import { formatCompactNumber, formatUsd } from '@/utils/usage';
+import authFileQuotaStyles from '@/pages/AuthFilesPage.module.scss';
+import styles from '../monitoring.module.scss';
+
+type AccountSummaryMetric = {
+  key: string;
+  label: string;
+  value: ReactNode;
+  valueClassName?: string;
+};
+
+const SuccessFailureValue = ({ success, failure }: { success: number; failure: number }) => (
+  <span className={styles.successFailureValue}>
+    <span className={styles.goodText}>{formatCompactNumber(success)}</span>
+    <span className={styles.badText}>({formatCompactNumber(failure)})</span>
+  </span>
+);
+
+const buildAccountCardFileName = (row: MonitoringAccountRow, quotaEntries: AccountQuotaEntry[] = []) => {
+  const quotaFileNames = Array.from(new Set(quotaEntries.map((entry) => entry.fileName).filter(Boolean)));
+  if (quotaFileNames.length > 0) return joinUnique(quotaFileNames, 1);
+
+  const fileName = row.authLabels.find((label) => label && label !== '-' && label.endsWith('.json'));
+  return fileName || row.authLabels.find((label) => label && label !== '-') || row.accountMasked || row.account;
+};
+
+const buildAccountCardProviderText = (row: MonitoringAccountRow) => {
+  const providers = row.providers.filter((provider) => provider && provider !== '-');
+  return providers.length > 0 ? joinUnique(providers, 2) : '-';
+};
+
+const sortAccountOverviewCardMetrics = (metrics: AccountSummaryMetric[], t: TFunction) => {
+  const labels: Record<string, string> = {
+    'total-tokens': t('monitoring.token_metric_total'),
+    'input-tokens': t('monitoring.token_metric_input'),
+    'output-tokens': t('monitoring.token_metric_output'),
+    'cached-tokens': t('monitoring.token_metric_cached'),
+  };
+  const order = ['total-tokens', 'input-tokens', 'output-tokens', 'cached-tokens'];
+  return order
+    .map((key) => {
+      const metric = metrics.find((item) => item.key === key);
+      return metric ? { ...metric, label: labels[key] } : undefined;
+    })
+    .filter(Boolean) as AccountSummaryMetric[];
+};
+
+const buildAccountSummaryMetrics = (
+  row: MonitoringAccountRow,
+  hasPrices: boolean,
+  locale: string,
+  t: TFunction
+): AccountSummaryMetric[] => [
+  {
+    key: 'total-calls',
+    label: t('monitoring.total_calls'),
+    value: formatCompactNumber(row.totalCalls),
+  },
+  {
+    key: 'success-calls',
+    label: t('monitoring.success_calls'),
+    value: <SuccessFailureValue success={row.successCalls} failure={row.failureCalls} />,
+  },
+  {
+    key: 'success-rate',
+    label: t('monitoring.call_success_rate'),
+    value: formatPercent(row.successRate),
+    valueClassName:
+      row.successRate >= 0.95
+        ? styles.goodText
+        : row.successRate >= 0.85
+          ? styles.warnText
+          : styles.badText,
+  },
+  {
+    key: 'total-tokens',
+    label: t('monitoring.total_tokens'),
+    value: formatCompactNumber(row.totalTokens),
+  },
+  {
+    key: 'input-tokens',
+    label: t('monitoring.input_tokens'),
+    value: formatCompactNumber(row.inputTokens),
+  },
+  {
+    key: 'output-tokens',
+    label: t('monitoring.output_tokens'),
+    value: formatCompactNumber(row.outputTokens),
+  },
+  {
+    key: 'cached-tokens',
+    label: t('monitoring.cached_tokens'),
+    value: formatCompactNumber(row.cachedTokens),
+  },
+  {
+    key: 'estimated-cost',
+    label: t('monitoring.estimated_cost'),
+    value: hasPrices ? formatUsd(row.totalCost) : '--',
+  },
+  {
+    key: 'latest-request-time',
+    label: t('monitoring.latest_request_time'),
+    value: new Date(row.lastSeenAt).toLocaleString(locale),
+  },
+];
+
+const ACCOUNT_QUOTA_RENDER_HELPERS: QuotaRenderHelpers = {
+  styles: {
+    ...authFileQuotaStyles,
+    quotaRow: `${authFileQuotaStyles.quotaRow} ${styles.accountQuotaRow}`,
+    quotaRowHeader: `${authFileQuotaStyles.quotaRowHeader} ${styles.accountQuotaRowHeader}`,
+    quotaModel: `${authFileQuotaStyles.quotaModel} ${styles.accountQuotaModel}`,
+    quotaMeta: `${authFileQuotaStyles.quotaMeta} ${styles.accountQuotaMeta}`,
+    quotaAmount: `${authFileQuotaStyles.quotaAmount} ${styles.accountQuotaAmount}`,
+    codexPlanValue: `${authFileQuotaStyles.codexPlanValue} ${styles.accountQuotaPlanValue}`,
+    premiumPlanValue: `${authFileQuotaStyles.premiumPlanValue} ${styles.accountQuotaPremiumPlanValue}`,
+    codexResetCreditRow: `${authFileQuotaStyles.codexResetCreditRow} ${styles.accountQuotaResetCreditRow}`,
+    codexResetCreditTime: `${authFileQuotaStyles.codexResetCreditTime} ${styles.accountQuotaResetCreditTime}`,
+  },
+  QuotaProgressBar: AuthFileQuotaProgressBar,
+};
+
+const getSuccessRateClassName = (rate: number) =>
+  rate >= 0.95 ? styles.goodText : rate >= 0.85 ? styles.warnText : styles.badText;
+
+const getAccountStatusDotClassName = (tone: AccountHealthTone) => {
+  if (tone === 'good') return styles.accountStatusDotEnabled;
+  if (tone === 'warn') return styles.accountStatusDotMixed;
+  return styles.accountStatusDotDisabled;
+};
+
+function AccountHealthStatusPanel({
+  row,
+  hasPrices,
+  locale,
+  t,
+  statusData,
+  scopeText,
+}: {
+  row: MonitoringAccountRow;
+  hasPrices: boolean;
+  locale: string;
+  t: TFunction;
+  statusData: AccountStatusData;
+  scopeText: string;
+}) {
+  const healthMetrics = [
+    { key: 'total-calls', label: t('monitoring.total_calls'), value: formatCompactNumber(row.totalCalls) },
+    {
+      key: 'success-failure',
+      label: t('monitoring.success_failure'),
+      value: <SuccessFailureValue success={row.successCalls} failure={row.failureCalls} />,
+    },
+    { key: 'estimated-cost', label: t('monitoring.estimated_cost'), value: hasPrices ? formatUsd(row.totalCost) : '--', className: styles.primaryText },
+    { key: 'success-rate', label: t('monitoring.success_rate'), value: formatPercent(row.successRate), className: getSuccessRateClassName(row.successRate) },
+  ];
+
+  return (
+    <section className={styles.accountOverviewStatusSection}>
+      <div className={styles.accountSectionHeader}>
+        <strong>{t('monitoring.account_health_status')}</strong>
+        <span className={styles.accountSectionInfo} title={t('monitoring.account_health_status_hint')}>
+          i
+        </span>
+      </div>
+      <div className={styles.healthMetricGrid}>
+        {healthMetrics.map((metric) => (
+          <div key={metric.key} className={styles.healthMetricItem}>
+            <span>{metric.label}</span>
+            <strong className={metric.className}>{metric.value}</strong>
+          </div>
+        ))}
+      </div>
+      <MonitoringHealthStatusBar statusData={statusData} locale={locale} t={t} showRate={false} />
+      <div className={styles.accountScopeText}>{scopeText}</div>
+    </section>
+  );
+}
+
+function AccountTokenMetricGrid({ metrics, t }: { metrics: AccountSummaryMetric[]; t: TFunction }) {
+  const getTokenMetricToneClassName = (key: string) => {
+    if (key === 'input-tokens') return styles.accountMetricIconInput;
+    if (key === 'output-tokens') return styles.accountMetricIconOutput;
+    if (key === 'cached-tokens') return styles.accountMetricIconCached;
+    return styles.accountMetricIconTotal;
+  };
+
+  return (
+    <section className={styles.accountTokenPanel}>
+      <div className={styles.accountSectionHeader}>
+        <strong>{t('monitoring.token_usage')}</strong>
+      </div>
+      <div className={styles.accountOverviewMetricGrid}>
+        {metrics.map((metric) => (
+          <div key={metric.key} className={styles.accountOverviewMetricCard}>
+            <span className={styles.accountOverviewMetricLabel}>
+              <span
+                className={[styles.accountMetricIcon, getTokenMetricToneClassName(metric.key)]
+                  .filter(Boolean)
+                  .join(' ')}
+                aria-hidden="true"
+              />
+              {metric.label}
+            </span>
+            <strong className={[styles.accountOverviewMetricValue, metric.valueClassName].filter(Boolean).join(' ')}>
+              {metric.value}
+            </strong>
+          </div>
+        ))}
+      </div>
+    </section>
+  );
+}
+
+function AccountModelUsageList({
+  row,
+  hasPrices,
+  locale,
+  t,
+  limit = 1,
+}: {
+  row: MonitoringAccountRow;
+  hasPrices: boolean;
+  locale: string;
+  t: TFunction;
+  limit?: number;
+}) {
+  const [showAll, setShowAll] = useState(false);
+  const [expandedModels, setExpandedModels] = useState<Record<string, boolean>>({});
+  const hasExtraModels = row.models.length > limit;
+  const visibleModels = showAll ? row.models : row.models.slice(0, limit);
+  const toggleModel = (key: string) => setExpandedModels((previous) => ({ ...previous, [key]: !previous[key] }));
+
+  return (
+    <section className={styles.accountModelListPanel}>
+      <div className={styles.accountSectionHeader}>
+        <strong>{t('monitoring.top_models')}</strong>
+        {hasExtraModels ? (
+          <button type="button" className={styles.accountModelViewAllButton} onClick={() => setShowAll((previous) => !previous)}>
+            {showAll ? t('monitoring.collapse') : t('monitoring.view_all')}
+          </button>
+        ) : null}
+      </div>
+
+      {visibleModels.length > 0 ? (
+        <div className={styles.accountModelList}>
+          {visibleModels.map((model) => {
+            const modelKey = `${row.id}-${model.model}`;
+            const isModelExpanded = Boolean(expandedModels[modelKey]);
+            return (
+              <div key={modelKey} className={styles.accountModelItem}>
+                <button
+                  type="button"
+                  className={styles.accountModelRow}
+                  onClick={() => toggleModel(modelKey)}
+                  aria-expanded={isModelExpanded}
+                >
+                  <span className={styles.accountModelName} title={model.model}>{model.model}</span>
+                  <span className={styles.accountModelMetaLine}>
+                    <span className={styles.accountModelStat}><small>{t('monitoring.ranking_metric_requests')}</small><strong>{formatCompactNumber(model.totalCalls)}</strong></span>
+                    <span className={styles.accountModelStat}><small>{t('monitoring.success_rate')}</small><strong className={getSuccessRateClassName(model.successRate)}>{formatPercent(model.successRate)}</strong></span>
+                    <span className={styles.accountModelStat}><small>{t('monitoring.ranking_metric_tokens')}</small><strong>{formatCompactNumber(model.totalTokens)}</strong></span>
+                    <span className={styles.accountModelStat}><small>{t('monitoring.ranking_metric_cost')}</small><strong>{hasPrices ? formatUsd(model.totalCost) : '--'}</strong></span>
+                    <span className={styles.accountModelChevron} aria-hidden="true">{isModelExpanded ? <IconChevronDown size={14} /> : '›'}</span>
+                  </span>
+                </button>
+                {isModelExpanded ? (
+                  <div className={styles.accountModelExpanded}>
+                    <div className={styles.accountModelExpandedItem}><small>{t('monitoring.input_tokens')}</small><strong>{formatCompactNumber(model.inputTokens)}</strong></div>
+                    <div className={styles.accountModelExpandedItem}><small>{t('monitoring.output_tokens')}</small><strong>{formatCompactNumber(model.outputTokens)}</strong></div>
+                    <div className={styles.accountModelExpandedItem}><small>{t('monitoring.cached_tokens')}</small><strong>{formatCompactNumber(model.cachedTokens)}</strong></div>
+                    <div className={styles.accountModelExpandedItem}><small>{t('monitoring.latest_request_time')}</small><strong>{new Date(model.lastSeenAt).toLocaleString(locale)}</strong></div>
+                  </div>
+                ) : null}
+              </div>
+            );
+          })}
+        </div>
+      ) : (
+        <div className={styles.emptyBlockSmall}>{t('monitoring.no_model_data')}</div>
+      )}
+    </section>
+  );
+}
+
+function AccountQuotaPanel({
+  quotaState,
+  quotaEntries,
+  locale,
+  t,
+  onRefreshQuota,
+}: {
+  quotaState?: AccountQuotaState;
+  quotaEntries: AccountQuotaEntry[];
+  locale: string;
+  t: TFunction;
+  onRefreshQuota: () => void;
+}) {
+  const quotaLoading = quotaState?.status === 'loading';
+  const lastQuotaSync = quotaState?.lastRefreshedAt && Number.isFinite(quotaState.lastRefreshedAt)
+    ? new Date(quotaState.lastRefreshedAt).toLocaleString(locale)
+    : '';
+  const quotaTitle = quotaEntries.length === 1 ? quotaEntries[0].providerLabel : t('quota_management.title');
+
+  const renderRefreshButton = () => (
+    <button type="button" className={styles.quotaRefreshButton} onClick={onRefreshQuota} disabled={quotaLoading}>
+      <IconRefreshCw size={14} className={quotaLoading ? styles.refreshIconSpinning : styles.refreshIcon} />
+      <span>{t('codex_quota.refresh_button')}</span>
+    </button>
+  );
+
+  return (
+    <section className={styles.quotaSection}>
+      <div className={styles.quotaSectionHeader}>
+        <div className={styles.quotaSectionTitleGroup}>
+          <strong>{quotaTitle}</strong>
+          {lastQuotaSync ? <span>{`${t('monitoring.last_sync')}: ${lastQuotaSync}`}</span> : null}
+        </div>
+        {renderRefreshButton()}
+      </div>
+
+      {quotaLoading && quotaEntries.length === 0 ? <div className={styles.quotaStateMessage}>{t('codex_quota.loading')}</div> : null}
+      {!quotaLoading && quotaState?.status === 'error' && quotaEntries.length === 0 ? (
+        <div className={styles.quotaStateMessage}>{t('codex_quota.load_failed', { message: quotaState.error || t('common.unknown_error') })}</div>
+      ) : null}
+      {!quotaLoading && quotaState?.status === 'success' && quotaEntries.length === 0 ? <div className={styles.quotaStateMessage}>{t('monitoring.account_quota_empty')}</div> : null}
+      {!quotaState && quotaEntries.length === 0 ? <div className={styles.quotaStateMessage}>{t('monitoring.account_quota_empty')}</div> : null}
+
+      {quotaEntries.length > 0 ? (
+        <div className={styles.quotaEntryGrid}>
+          {quotaEntries.map((entry) => (
+            <div key={entry.key} className={styles.quotaEntryCard}>
+              <div className={styles.quotaEntryHeader}>
+                <div className={styles.quotaEntryMain}>
+                  <strong>{entry.authLabel}</strong>
+                </div>
+              </div>
+
+              {entry.quota?.status === 'loading' ? (
+                <div className={styles.quotaStateMessage}>{t(`${entry.config.i18nPrefix}.loading`)}</div>
+              ) : entry.quota?.status === 'error' ? (
+                <div className={styles.quotaStateMessage}>
+                  {t(`${entry.config.i18nPrefix}.load_failed`, { message: resolveQuotaErrorMessage(t, entry.quota) })}
+                </div>
+              ) : hasUsableQuotaContent(entry.quota) ? (
+                <div className={`${authFileQuotaStyles.quotaSection} ${styles.accountQuotaContent}`}>
+                  {entry.config.renderQuotaItems(entry.quota!, t, ACCOUNT_QUOTA_RENDER_HELPERS)}
+                </div>
+              ) : (
+                <div className={styles.quotaStateMessage}>{t(`${entry.config.i18nPrefix}.idle`)}</div>
+              )}
+            </div>
+          ))}
+        </div>
+      ) : null}
+    </section>
+  );
+}
+
+function AccountExpandedDetails({
+  row,
+  hasPrices,
+  locale,
+  t,
+  quotaState,
+  quotaEntries,
+  onRefreshQuota,
+}: {
+  row: MonitoringAccountRow;
+  hasPrices: boolean;
+  locale: string;
+  t: TFunction;
+  quotaState?: AccountQuotaState;
+  quotaEntries: AccountQuotaEntry[];
+  onRefreshQuota: () => void;
+}) {
+  return (
+    <div className={styles.accountOverviewCardBody}>
+      <AccountQuotaPanel quotaState={quotaState} quotaEntries={quotaEntries} locale={locale} t={t} onRefreshQuota={onRefreshQuota} />
+      <AccountModelUsageList row={row} hasPrices={hasPrices} locale={locale} t={t} />
+    </div>
+  );
+}
+
+export function AccountOverviewCard({
+  row,
+  hasPrices,
+  locale,
+  t,
+  isExpanded,
+  statusData,
+  scopeText,
+  quotaState,
+  quotaEntries,
+  onToggle,
+  onRefreshQuota,
+}: {
+  row: MonitoringAccountRow;
+  hasPrices: boolean;
+  locale: string;
+  t: TFunction;
+  isExpanded: boolean;
+  statusData: AccountStatusData;
+  scopeText: string;
+  quotaState?: AccountQuotaState;
+  quotaEntries: AccountQuotaEntry[];
+  onToggle: () => void;
+  onRefreshQuota: () => void;
+}) {
+  const summaryMetrics = buildAccountSummaryMetrics(row, hasPrices, locale, t);
+  const cardMetrics = sortAccountOverviewCardMetrics(summaryMetrics, t);
+  const tone = getAccountHealthTone(row);
+  const latestRequestText = new Date(row.lastSeenAt).toLocaleString(locale);
+  const accountLabel = buildAccountCardFileName(row, quotaEntries);
+  const providerText = buildAccountCardProviderText(row);
+
+  return (
+    <Card
+      className={[
+        styles.accountOverviewCard,
+        isExpanded ? styles.accountOverviewCardExpanded : '',
+      ]
+        .filter(Boolean)
+        .join(' ')}
+    >
+      <div className={styles.accountOverviewCardHeader}>
+        <div className={styles.accountTitleRow}>
+          <button
+            type="button"
+            className={styles.accountButton}
+            onClick={onToggle}
+            aria-expanded={isExpanded}
+            title={accountLabel}
+          >
+            <span className={styles.accountExpandGlyph} aria-hidden="true">
+              {isExpanded ? <IconChevronUp size={15} /> : <IconChevronDown size={15} />}
+            </span>
+            <span className={styles.accountIdentityLine}>
+              <span className={[styles.accountStatusDot, getAccountStatusDotClassName(tone)].filter(Boolean).join(' ')} aria-hidden="true" />
+              <span className={styles.accountButtonLabel}>{accountLabel}</span>
+            </span>
+          </button>
+          <span className={`${styles.accountHealthBadge} ${styles[`accountHealthBadge${tone}`]}`}>
+            {tone === 'good' ? t('monitoring.health_good') : tone === 'warn' ? t('monitoring.health_warn') : t('monitoring.health_bad')}
+          </span>
+        </div>
+        <div className={styles.accountMetaRow}>
+          <span className={styles.accountOverviewCardTimestamp} title={providerText}>{providerText}</span>
+          <span className={styles.accountMetaSeparator}>·</span>
+          <span className={styles.accountOverviewCardTimestamp}>{t('monitoring.latest_request_time_value', { value: latestRequestText })}</span>
+        </div>
+      </div>
+
+      <AccountHealthStatusPanel row={row} hasPrices={hasPrices} locale={locale} t={t} statusData={statusData} scopeText={scopeText} />
+      <AccountTokenMetricGrid metrics={cardMetrics} t={t} />
+
+      {isExpanded ? (
+        <AccountExpandedDetails
+          row={row}
+          hasPrices={hasPrices}
+          locale={locale}
+          t={t}
+          quotaState={quotaState}
+          quotaEntries={quotaEntries}
+          onRefreshQuota={onRefreshQuota}
+        />
+      ) : null}
+    </Card>
+  );
+}
