+++ cliproxyapi-pro-management/tests/test_monitoring_module_boundaries.py
@@ -10,13 +10,16 @@ class MonitoringModuleBoundariesTest(unittest.TestCase):
     def test_page_remains_a_controller_instead_of_reabsorbing_extracted_logic(self) -> None:
         source = PAGE_PATH.read_text()
 
-        self.assertLess(len(source.splitlines()), 4000)
+        self.assertLess(len(source.splitlines()), 2300)
         self.assertNotIn('const buildUsageTrendAnalytics =', source)
         self.assertNotIn('const buildRealtimeLogPageRows =', source)
         self.assertNotIn('const normalizeRealtimeLogColumns =', source)
         self.assertNotIn('function RealtimeCostCell(', source)
         self.assertNotIn('function UsageTrendPanel(', source)
         self.assertNotIn('function MonitoringHealthStatusBar(', source)
+        self.assertNotIn('function AccountStatsPanel(', source)
+        self.assertNotIn('open={isMonitoringSettingsOpen}', source)
+        self.assertNotIn('open={isPriceModalOpen}', source)
 
     def test_extracted_modules_own_their_domain_contracts(self) -> None:
         analytics = (OVERLAY_ROOT / 'features/monitoring/monitoringAnalytics.ts').read_text()
@@ -29,6 +32,13 @@ class MonitoringModuleBoundariesTest(unittest.TestCase):
         self.assertIn('export type RealtimeLogRow', realtime)
         self.assertIn('export const normalizeRealtimeLogColumns', preferences)
         self.assertIn('export const buildAccountStatusData', health)
+
+    def test_monitoring_feature_owns_shared_styles(self) -> None:
+        feature_root = OVERLAY_ROOT / 'features/monitoring'
+        self.assertTrue((feature_root / 'monitoring.module.scss').is_file())
+        self.assertFalse((OVERLAY_ROOT / 'pages/MonitoringCenterPage.module.scss').exists())
+        for component in (feature_root / 'components').glob('*.tsx'):
+            self.assertNotIn("@/pages/MonitoringCenterPage.module.scss", component.read_text())
 
 
 if __name__ == '__main__':
